Changed average so that modules depending on average.c need to declare their own buffer size instead

+++ tests/test_circular_buffer.c
@@ -188,7 +188,7 @@ void test_buf_is_signed(void)
 {
 
     writeCircBuf(&buff, 0);
-    int16_t entry;
+    int32_t entry;
     readCircBuf(&buff, &entry);
     TEST_ASSERT_EQUAL(-1, entry-1);
 }
